Started spliting the Command Type and Status Type out of the Navigation class. Still needs work to complete the split. Working towards an inhertance based version.
