+++ DataDictionary.Main/Enumerations/Navigation.Enumeration.cs
@@ -12,11 +12,6 @@ namespace DataDictionary.Main.Enumerations
         Icon WindowIcon { get; }
 
         /// <summary>
-        /// List of Images for the Scope Type assocated with Commands.
-        /// </summary>
-        IReadOnlyDictionary<CommandType, Func<Image>> CommandImages { get; }
-
-        /// <summary>
         /// Grouping behavior.
         /// When True, items with the same ScopeType will be group together.
         /// When False, items will not be group together and appear as individual entries.
@@ -44,40 +39,6 @@ namespace DataDictionary.Main.Enumerations
             /// <inheritdoc/>
             public Icon WindowIcon { get; init; } = Resources.Icon_UnknownMember;
             static readonly Icon defaultIcon = Resources.Icon_UnknownMember;
-
-            // List of default overlay Images
-            static readonly Dictionary<CommandType, Image> commandOverlay = new Dictionary<CommandType, Image>()
-            {
-                {CommandType.Browse,  Resources.ItemBrowse},
-                {CommandType.Add,     Resources.ItemNew},
-                {CommandType.Open,    Resources.ItemOpen},
-                {CommandType.Save,    Resources.ItemSave},
-                {CommandType.Delete,  Resources.ItemDelete},
-                {CommandType.Export,  Resources.ItemExport},
-                {CommandType.Import,  Resources.ItemImport},
-                {CommandType.Refresh, Resources.ItemRefresh},
-                {CommandType.Select,  Resources.ItemSelect},
-                {CommandType.Sync,    Resources.ItemSync},
-            };
-
-            /// <inheritdoc/>
-            public IReadOnlyDictionary<CommandType, Func<Image>> CommandImages
-            { get { return commandImages; } }
-            Dictionary<CommandType, Func<Image>> commandImages { get; init; } = new Dictionary<CommandType, Func<Image>>();
-
-            static readonly Dictionary<StatusType, Image> statusOverlay = new Dictionary<StatusType, Image>()
-            {
-                {StatusType.Ok, Resources.StatusOK},
-                {StatusType.Error, Resources.StatusError},
-                {StatusType.Information, Resources.StatusInformation},
-                {StatusType.Invalid, Resources.StatusInvalid},
-                {StatusType.No, Resources.StatusNo},
-            };
-
-            public IReadOnlyDictionary<StatusType, Func<Image>> StatusImages
-            { get { return statusImages; } }
-            Dictionary<StatusType, Func<Image>> statusImages { get; init; } = new Dictionary<StatusType, Func<Image>>();
-
 
             /// <inheritdoc/>
             public Boolean GroupBy { get; init; } = true;
